Creating repository for collection of cgi files
This repository will be for me to practice creating cgi scripts.  Most
of these will be small one-of things.

As of right now, they all work except for the one that tries display
the local time.  Also the one that tries to display the list of processes
is not finished.